# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRestorePointResponse": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "is_idempotent": {
            "title": "Is Idempotent",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "project_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Uuid"
          },
          "queue_position": {
            "title": "Queue Position",
            "type": "integer"
          },
          "restore_point_id": {
            "title": "Restore Point Id",
            "type": "string"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Uuid"
          },
          "total_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Steps"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "job_id",
          "project",
          "project_uuid",
          "tenant_uuid",
          "created_by",
          "action",
          "status",
          "message",
          "progress",
          "current_step",
          "total_steps",
          "started_at",
          "finished_at",
          "error_code",
          "is_idempotent",
          "retryable",
          "retry_of",
          "attempt",
          "created_at",
          "updated_at",
          "queue_position",
          "restore_point_id"
        ],
        "title": "CreateRestorePointResponse",
        "type": "object"
      },
...
      "DeleteRestorePointResponse": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "is_idempotent": {
            "title": "Is Idempotent",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "project_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Uuid"
          },
          "queue_position": {
            "title": "Queue Position",
            "type": "integer"
          },
          "restore_point_id": {
            "title": "Restore Point Id",
            "type": "string"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Uuid"
          },
          "total_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Steps"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "job_id",
          "project",
          "project_uuid",
          "tenant_uuid",
          "created_by",
          "action",
          "status",
          "message",
          "progress",
          "current_step",
          "total_steps",
          "started_at",
          "finished_at",
          "error_code",
          "is_idempotent",
          "retryable",
          "retry_of",
          "attempt",
          "created_at",
          "updated_at",
          "queue_position",
          "restore_point_id"
        ],
        "title": "DeleteRestorePointResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListRestorePointsResponse": {
        "additionalProperties": true,
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "permissions": {
            "$ref": "#/components/schemas/RestorePointsPermissions"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RestorePointItem"
            },
            "title": "Points",
            "type": "array"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "required": [
          "project",
          "limit",
          "permissions",
          "points"
        ],
        "title": "ListRestorePointsResponse",
        "type": "object"
      },
...
      "RestorePointCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "RestorePointCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane for isolated Supabase projects on shared infrastructure.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Supabase Multitenant Projects API",
    "version": "0.13.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_name}/restore-points": {
      "get": {
        "operationId": "list_project_restore_points_api_projects__project_name__restore_points_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRestorePointsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Restore Points",
        "tags": [
          "restore-points"
        ]
      },
      "post": {
        "operationId": "create_project_restore_point_api_projects__project_name__restore_points_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePointCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRestorePointResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project Restore Point",
        "tags": [
          "restore-points"
        ]
      }
    },
    "/api/projects/{project_name}/restore-points/{point_id}": {
      "delete": {
        "operationId": "delete_project_restore_point_api_projects__project_name__restore_points__point_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "point_id",
            "required": true,
            "schema": {
              "title": "Point Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRestorePointResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project Restore Point",
        "tags": [
          "restore-points"
        ]
      }
    },
    "/api/projects/{project_name}/restore-points/{point_id}/restore": {
      "post": {
        "operationId": "restore_project_restore_point_api_projects__project_name__restore_points__point_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "point_id",
            "required": true,
            "schema": {
              "title": "Point Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreRestorePointResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Project Restore Point",
        "tags": [
          "restore-points"
        ]
      }
    },
...
  }
}